+++ clients/crypto-loan/src/main/java/com/binance/connector/client/crypto_loan/rest/api/CryptoLoanRestApi.java
@@ -7,23 +7,12 @@ import com.binance.connector.client.common.configuration.ClientConfiguration;
 import com.binance.connector.client.crypto_loan.rest.CryptoLoanRestApiUtil;
 import com.binance.connector.client.crypto_loan.rest.model.CheckCollateralRepayRateResponse;
 import com.binance.connector.client.crypto_loan.rest.model.CheckCollateralRepayRateStableRateResponse;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanAdjustLtvRequest;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanAdjustLtvResponse;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanBorrowRequest;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanBorrowResponse;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanCustomizeMarginCallRequest;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanCustomizeMarginCallResponse;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanRepayRequest;
-import com.binance.connector.client.crypto_loan.rest.model.CryptoLoanRepayResponse;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanAdjustLtvRequest;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanAdjustLtvResponse;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanBorrowRequest;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanBorrowResponse;
-import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanCollateralRepaymentRequest;
-import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanCollateralRepaymentResponse;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanRepayRequest;
 import com.binance.connector.client.crypto_loan.rest.model.FlexibleLoanRepayResponse;
-import com.binance.connector.client.crypto_loan.rest.model.GetCollateralAssetsDataResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetCryptoLoansIncomeHistoryResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetFlexibleLoanAssetsDataResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetFlexibleLoanBorrowHistoryResponse;
@@ -34,9 +23,7 @@ import com.binance.connector.client.crypto_loan.rest.model.GetFlexibleLoanOngoin
 import com.binance.connector.client.crypto_loan.rest.model.GetFlexibleLoanRepaymentHistoryResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetLoanBorrowHistoryResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetLoanLtvAdjustmentHistoryResponse;
-import com.binance.connector.client.crypto_loan.rest.model.GetLoanOngoingOrdersResponse;
 import com.binance.connector.client.crypto_loan.rest.model.GetLoanRepaymentHistoryResponse;
-import com.binance.connector.client.crypto_loan.rest.model.GetLoanableAssetsDataResponse;
 
 public class CryptoLoanRestApi {
 
@@ -123,31 +110,6 @@ public class CryptoLoanRestApi {
     public ApiResponse<FlexibleLoanBorrowResponse> flexibleLoanBorrow(
             FlexibleLoanBorrowRequest flexibleLoanBorrowRequest) throws ApiException {
         return flexibleRateApi.flexibleLoanBorrow(flexibleLoanBorrowRequest);
-    }
-
-    /**
-     * Flexible Loan Collateral Repayment (TRADE) Weight: 6000
-     *
-     * @param flexibleLoanCollateralRepaymentRequest (required)
-     * @return ApiResponse&lt;FlexibleLoanCollateralRepaymentResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Flexible Loan Collateral Repayment </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/flexible-rate/trade/Flexible-Loan-Collateral-Repay">Flexible
-     *     Loan Collateral Repayment (TRADE) Documentation</a>
-     */
-    public ApiResponse<FlexibleLoanCollateralRepaymentResponse> flexibleLoanCollateralRepayment(
-            FlexibleLoanCollateralRepaymentRequest flexibleLoanCollateralRepaymentRequest)
-            throws ApiException {
-        return flexibleRateApi.flexibleLoanCollateralRepayment(
-                flexibleLoanCollateralRepaymentRequest);
     }
 
     /**
@@ -441,126 +403,6 @@ public class CryptoLoanRestApi {
     }
 
     /**
-     * Crypto Loan Adjust LTV(TRADE) Crypto Loan Adjust LTV Weight: 6000
-     *
-     * @param cryptoLoanAdjustLtvRequest (required)
-     * @return ApiResponse&lt;CryptoLoanAdjustLtvResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Crypto Loan Adjust LTV </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/trade/Crypto-Loan-Adjust-LTV">Crypto
-     *     Loan Adjust LTV(TRADE) Documentation</a>
-     */
-    public ApiResponse<CryptoLoanAdjustLtvResponse> cryptoLoanAdjustLtv(
-            CryptoLoanAdjustLtvRequest cryptoLoanAdjustLtvRequest) throws ApiException {
-        return stableRateApi.cryptoLoanAdjustLtv(cryptoLoanAdjustLtvRequest);
-    }
-
-    /**
-     * Crypto Loan Borrow(TRADE) Crypto Loan Borrow Weight: 6000
-     *
-     * @param cryptoLoanBorrowRequest (required)
-     * @return ApiResponse&lt;CryptoLoanBorrowResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Crypto Loan Borrow </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/trade/Crypto-Loan-Borrow">Crypto
-     *     Loan Borrow(TRADE) Documentation</a>
-     */
-    public ApiResponse<CryptoLoanBorrowResponse> cryptoLoanBorrow(
-            CryptoLoanBorrowRequest cryptoLoanBorrowRequest) throws ApiException {
-        return stableRateApi.cryptoLoanBorrow(cryptoLoanBorrowRequest);
-    }
-
-    /**
-     * Crypto Loan Customize Margin Call(TRADE) Customize Margin Call Weight: 6000
-     *
-     * @param cryptoLoanCustomizeMarginCallRequest (required)
-     * @return ApiResponse&lt;CryptoLoanCustomizeMarginCallResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Crypto Loan Customize Margin Call </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/trade/Crypto-Loan-Customize-Margin-Call">Crypto
-     *     Loan Customize Margin Call(TRADE) Documentation</a>
-     */
-    public ApiResponse<CryptoLoanCustomizeMarginCallResponse> cryptoLoanCustomizeMarginCall(
-            CryptoLoanCustomizeMarginCallRequest cryptoLoanCustomizeMarginCallRequest)
-            throws ApiException {
-        return stableRateApi.cryptoLoanCustomizeMarginCall(cryptoLoanCustomizeMarginCallRequest);
-    }
-
-    /**
-     * Crypto Loan Repay(TRADE) Crypto Loan Repay Weight: 6000
-     *
-     * @param cryptoLoanRepayRequest (required)
-     * @return ApiResponse&lt;CryptoLoanRepayResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Crypto Loan Repay </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/trade/Crypto-Loan-Repay">Crypto
-     *     Loan Repay(TRADE) Documentation</a>
-     */
-    public ApiResponse<CryptoLoanRepayResponse> cryptoLoanRepay(
-            CryptoLoanRepayRequest cryptoLoanRepayRequest) throws ApiException {
-        return stableRateApi.cryptoLoanRepay(cryptoLoanRepayRequest);
-    }
-
-    /**
-     * Get Collateral Assets Data(USER_DATA) Get LTV information and collateral limit of collateral
-     * assets. The collateral limit is shown in USD value. Weight: 400
-     *
-     * @param collateralCoin (optional)
-     * @param vipLevel Default: user&#39;s vip level. Send \&quot;-1\&quot; to check specified
-     *     configuration (optional)
-     * @param recvWindow (optional)
-     * @return ApiResponse&lt;GetCollateralAssetsDataResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Get Collateral Assets Data </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/market-data/Get-Collateral-Assets-Data">Get
-     *     Collateral Assets Data(USER_DATA) Documentation</a>
-     */
-    public ApiResponse<GetCollateralAssetsDataResponse> getCollateralAssetsData(
-            String collateralCoin, Long vipLevel, Long recvWindow) throws ApiException {
-        return stableRateApi.getCollateralAssetsData(collateralCoin, vipLevel, recvWindow);
-    }
-
-    /**
      * Get Crypto Loans Income History(USER_DATA) Get Crypto Loans Income History * If startTime and
      * endTime are not sent, the recent 7-day data will be returned. * The max interval between
      * startTime and endTime is 30 days. Weight: 6000
@@ -590,7 +432,7 @@ public class CryptoLoanRestApi {
      *     Crypto Loans Income History(USER_DATA) Documentation</a>
      */
     public ApiResponse<GetCryptoLoansIncomeHistoryResponse> getCryptoLoansIncomeHistory(
-            String asset, Long type, Long startTime, Long endTime, Long limit, Long recvWindow)
+            String asset, String type, Long startTime, Long endTime, Long limit, Long recvWindow)
             throws ApiException {
         return stableRateApi.getCryptoLoansIncomeHistory(
                 asset, type, startTime, endTime, limit, recvWindow);
@@ -679,41 +521,6 @@ public class CryptoLoanRestApi {
     }
 
     /**
-     * Get Loan Ongoing Orders(USER_DATA) Get Loan Ongoing Orders Weight: 300
-     *
-     * @param orderId orderId in &#x60;POST /sapi/v1/loan/borrow&#x60; (optional)
-     * @param loanCoin (optional)
-     * @param collateralCoin (optional)
-     * @param current Current querying page. Start from 1; default: 1; max: 1000 (optional)
-     * @param limit Default: 10; max: 100 (optional)
-     * @param recvWindow (optional)
-     * @return ApiResponse&lt;GetLoanOngoingOrdersResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Get Loan Ongoing Orders </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/user-information/Get-Loan-Ongoing-Orders">Get
-     *     Loan Ongoing Orders(USER_DATA) Documentation</a>
-     */
-    public ApiResponse<GetLoanOngoingOrdersResponse> getLoanOngoingOrders(
-            Long orderId,
-            String loanCoin,
-            String collateralCoin,
-            Long current,
-            Long limit,
-            Long recvWindow)
-            throws ApiException {
-        return stableRateApi.getLoanOngoingOrders(
-                orderId, loanCoin, collateralCoin, current, limit, recvWindow);
-    }
-
-    /**
      * Get Loan Repayment History(USER_DATA) Get Loan Repayment History * If startTime and endTime
      * are not sent, the recent 90-day data will be returned. * The max interval between startTime
      * and endTime is 180 days. Weight: 400
@@ -752,32 +559,5 @@ public class CryptoLoanRestApi {
             throws ApiException {
         return stableRateApi.getLoanRepaymentHistory(
                 orderId, loanCoin, collateralCoin, startTime, endTime, current, limit, recvWindow);
-    }
-
-    /**
-     * Get Loanable Assets Data(USER_DATA) Get interest rate and borrow limit of loanable assets.
-     * The borrow limit is shown in USD value. Weight: 400
-     *
-     * @param loanCoin (optional)
-     * @param vipLevel Default: user&#39;s vip level. Send \&quot;-1\&quot; to check specified
-     *     configuration (optional)
-     * @param recvWindow (optional)
-     * @return ApiResponse&lt;GetLoanableAssetsDataResponse&gt;
-     * @throws ApiException If fail to call the API, e.g. server error or cannot deserialize the
-     *     response body
-     * @http.response.details
-     *     <table border="1">
-     * <caption>Response Details</caption>
-     * <tr><td> Status Code </td><td> Description </td><td> Response Headers </td></tr>
-     * <tr><td> 200 </td><td> Get Loanable Assets Data </td><td>  -  </td></tr>
-     * </table>
-     *
-     * @see <a
-     *     href="https://developers.binance.com/docs/crypto_loan/stable-rate/market-data/Get-Loanable-Assets-Data">Get
-     *     Loanable Assets Data(USER_DATA) Documentation</a>
-     */
-    public ApiResponse<GetLoanableAssetsDataResponse> getLoanableAssetsData(
-            String loanCoin, Long vipLevel, Long recvWindow) throws ApiException {
-        return stableRateApi.getLoanableAssetsData(loanCoin, vipLevel, recvWindow);
     }
 }