+++ clients/derivatives-trading-usds-futures/src/main/java/com/binance/connector/client/derivatives_trading_usds_futures/rest/api/TradeApi.java
@@ -83,7 +83,7 @@ public class TradeApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-derivatives-trading-usds-futures/1.2.0 (Java/%s; %s; %s)",
+                    "binance-derivatives-trading-usds-futures/2.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;
 
@@ -2804,8 +2804,7 @@ public class TradeApi {
      * the amendment in the following situations: * when the order is in partially filled status and
      * the new &#x60;quantity&#x60; &lt;&#x3D; &#x60;executedQty&#x60; * When the order is
      * &#x60;GTX&#x60; and the new price will cause it to be executed immediately * One order can
-     * only be modfied for less than 10000 times * Modify order will set
-     * &#x60;selfTradePreventionMode&#x60; to &#x60;NONE&#x60; Weight: 1 on 10s order rate
+     * only be modfied for less than 10000 times Weight: 1 on 10s order rate
      * limit(X-MBX-ORDER-COUNT-10S); 1 on 1min order rate limit(X-MBX-ORDER-COUNT-1M); 1 on IP rate
      * limit(x-mbx-used-weight-1m)
      *