+++ clients/mining/src/main/java/com/binance/connector/client/mining/rest/api/MiningApi.java
@@ -57,7 +57,7 @@ public class MiningApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-mining/1.1.0 (Java/%s; %s; %s)",
+                    "binance-mining/2.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;
 
@@ -1341,7 +1341,11 @@ public class MiningApi {
             localVarHeaderParams.put("Content-Type", localVarContentType);
         }
         List<String> localVarAuthNames = new ArrayList<>();
-        localVarAuthNames.addAll(Arrays.asList(new String[] {}));
+        localVarAuthNames.addAll(
+                Arrays.asList(
+                        new String[] {
+                            "binanceSignature",
+                        }));
         if (HAS_TIME_UNIT) {
             localVarAuthNames.add("timeUnit");
         }