+++ clients/sub-account/src/main/java/com/binance/connector/client/sub_account/rest/api/AccountManagementApi.java
@@ -57,7 +57,7 @@ public class AccountManagementApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-sub-account/1.1.0 (Java/%s; %s; %s)",
+                    "binance-sub-account/2.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;