+++ clients/crypto-loan/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 2.0.0 - 2025-06-02
+
+### Removed (8)
+
+- `flexibleLoanCollateralRepayment()` (`POST /sapi/v2/loan/flexible/repay/collateral`)
+- `cryptoLoanAdjustLtv()` (`POST /sapi/v1/loan/adjust/ltv`)
+- `cryptoLoanBorrow()` (`POST /sapi/v1/loan/borrow`)
+- `cryptoLoanCustomizeMarginCall()` (`POST /sapi/v1/loan/customize/margin_call`)
+- `cryptoLoanRepay()` (`POST /sapi/v1/loan/repay`)
+- `getCollateralAssetsData()` (`GET /sapi/v1/loan/collateral/data`)
+- `getLoanOngoingOrders()` (`GET /sapi/v1/loan/ongoing/orders`)
+- `getLoanableAssetsData()` (`GET /sapi/v1/loan/loanable/data`)
+
 ## 1.1.0 - 2025-05-02
 
 - Update `binance/common` module to version `1.1.0`.