+++ clients/simple-earn/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 2.0.0 - 2025-05-28
+
+### Changed (2)
+
+- Added parameter `recvWindow`
+  - affected methods:
+    - `getFlexibleRedemptionRecord()` (`GET /sapi/v1/simple-earn/flexible/history/redemptionRecord`)
+- Marked as signed the following endpoints:
+  - `GET /sapi/v1/simple-earn/locked/position`
+
 ## 1.1.0 - 2025-05-02
 
 - Update `binance/common` module to version `1.1.0`.