+++ clients/sub-account/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.0 - 2025-05-26
+
+### Changed (2)
+
+- Modified parameter `ipAddress` for the following methods:
+  - `addIpRestrictionForSubAccountApiKey()` (`POST /sapi/v2/sub-account/subAccountApi/ipRestriction`)
+  - `deleteIpListForASubAccountApiKey()` (`DELETE /sapi/v1/sub-account/subAccountApi/ipRestriction/ipList`)
+- Fields `subUserId` and `remark` added to the response of `/sapi/v1/sub-account/list`
+
 ## 1.1.0 - 2025-05-02
 
 - Update `binance/common` module to version `1.1.0`.